requirements.txt
# Requirements for deep-research-agent
pydantic>=2.6,<3
python-dotenv>=1.0
sendgrid>=6.11
gradio>=4.0
openai>=1.0.0
openai-agents>=0.0.10
pytest>=8.0
pytest-asyncio>=0.23
ruff>=0.5
markdown>=3.0
python-dateutil>=2.8.0
bleach>=6.0.0
hypothesis>=6.0.0
